+++ betterdesk.sh
@@ -2587,6 +2587,7 @@ Type=simple
 User=root
 WorkingDirectory=$RUSTDESK_PATH
 Environment=AUTH_DB_PATH=$CONSOLE_PATH/data/auth.db
+Environment=MESH_ENABLED=Y
 $CONNECTION_MODE_ENV_BLOCK
 ExecStart=$RUSTDESK_PATH/betterdesk-server -mode all -relay-servers $server_ip $systemd_db_arg -key-file $RUSTDESK_PATH/id_ed25519 -api-port $API_PORT -signal-rate-limit-per-ip $signal_rate_limit $init_admin_arg $tls_arg
 Restart=always
@@ -2990,6 +2991,7 @@ ExecStart=$GO_BINARY_PATH $SERVER_ARGS
 Restart=always
 RestartSec=5
 $GO_ENV
+Environment=MESH_ENABLED=Y
 $CONNECTION_MODE_ENV_BLOCK
 
 # Hardening